move out add_contact into a dedicated file
